小数位称重

+++ src/main/java/mob/code/supermarket/controller/QuantifiedBarcode.java
@@ -3,27 +3,27 @@ package mob.code.supermarket.controller;
 public class QuantifiedBarcode {
 
     private final String code;
-    private int quantity;
+    private double quantity;
 
     public QuantifiedBarcode(String scan) {
         code = scan.split("-")[0];
         String[] split = scan.split("-");
         quantity = 1;
         if (split.length > 1) {
-            quantity = Integer.parseInt(split[1]);
+            quantity = Double.parseDouble(split[1]);
         }
     }
 
-    public QuantifiedBarcode(String code, Long quantity) {
+    public QuantifiedBarcode(String code, double quantity) {
         this.code = code;
-        this.quantity = quantity.intValue();
+        this.quantity = quantity;
     }
 
     public String getBarCode() {
         return code;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 }
